+++ bocadillo/__init__.py
@@ -12,6 +12,7 @@ from .response import Response
 from .sse import server_event
 from .staticfiles import static
 from .templates import Templates
+from .testing import create_client, LiveServer
 from .views import view
 from .websockets import WebSocket, WebSocketDisconnect