+++ tests/test_cors.py
@@ -1,5 +1,4 @@
-from bocadillo import configure
-from bocadillo.testing import create_client
+from bocadillo import configure, create_client
 
 
 def test_no_allowed_origins_by_default(raw_app):